make some refactor to the design system, like reduce the files used and unnecessary same name files like app-theme and tudee-theme, make the colors together in the same data class body for easy to switch between themes type safety and good structure
