+++ app/src/main/java/com/washingtondcsquad/tudee/presentation/design/AppTheme.kt
@@ -1,0 +1,49 @@
+package com.washingtondcsquad.tudee.presentation.design
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.staticCompositionLocalOf
+import com.washingtondcsquad.tudee.presentation.design.textStyle.TudeeTextStyle
+import com.washingtondcsquad.tudee.presentation.design.textStyle.defaultTextStyle
+
+@Stable
+object AppTheme {
+
+    val textStyle: TudeeTextStyle
+        @Composable
+        @ReadOnlyComposable
+        get()= LocalTudeeTextStyle.current
+
+    val colors: TudeeColors
+        @Composable
+        @ReadOnlyComposable
+        get() = LocalTudeeColors.current
+
+    @Composable
+    operator fun invoke(
+        useDarkTheme: Boolean,
+        content: @Composable () -> Unit
+    ){
+        val colors = if (!useDarkTheme) {
+            TudeeColors.lightTudeeColors
+        } else {
+            TudeeColors.darkTudeeColors
+        }
+
+        CompositionLocalProvider(
+            LocalTudeeTextStyle provides defaultTextStyle,
+            LocalTudeeColors provides colors
+        ) {
+            content()
+        }
+
+    }
+
+    private val LocalTudeeColors = staticCompositionLocalOf { TudeeColors.lightTudeeColors }
+    private val LocalTudeeTextStyle = staticCompositionLocalOf { defaultTextStyle }
+}
+
+
+
