--- app/src/main/java/com/washingtondcsquad/tudee/presentation/design/colors/TudeeColors.kt
@@ -1,35 +1,0 @@
-package com.washingtondcsquad.tudee.presentation.design.colors
-
-import androidx.compose.runtime.Immutable
-import androidx.compose.ui.graphics.Color
-
-@Immutable
-data class TudeeColors(
-    val primary: Color,
-    val secondary: Color,
-    val primaryVarient: Color,
-    val primaryGradient: List<Color>,
-    val title: Color,
-    val body: Color,
-    val hint: Color,
-    val stroke: Color,
-    val surfaceLow: Color,
-    val surface: Color,
-    val surfaceHigh: Color,
-    val onPrimary: Color,
-    val onPrimaryCaption: Color,
-    val onPrimaryCard: Color,
-    val onPrimaryStroke: Color,
-    val disable: Color,
-    val pinkAccent: Color,
-    val yellowAccent: Color,
-    val greenAccent: Color,
-    val purpleAccent: Color,
-    val error: Color,
-    val overlay: Color,
-    val emojiTint: Color,
-    val yellowVariant: Color,
-    val greenVariant: Color,
-    val purpleVariant: Color,
-    val errorVariant: Color,
-)
