--- app/src/main/java/com/washingtondcsquad/tudee/presentation/design/textStyle/font.kt
@@ -1,12 +1,0 @@
-package com.washingtondcsquad.tudee.presentation.design.textStyle
-
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import com.washingtondcsquad.tudee.R
-
-val nunito = FontFamily(
-    Font(R.font.nunito, weight = FontWeight.Normal),
-    Font(R.font.nunito_medium, weight = FontWeight.Medium),
-    Font(R.font.nunito_semibold, weight = FontWeight.SemiBold),
-)
